Go to next and previous phase, fixes, seconds in DB etc

+++ trunk/src/businessLogic/insertNewGame.php
@@ -5,31 +5,32 @@ require("./databaseLogin.php");
 if( $_SESSION['loggedIn'] == "yes" and
 	$_SESSION['role'] == "organizer"){
 	
-	$length1a	= mysql_real_escape_string($_POST['length1a']);
-	$length1b	= mysql_real_escape_string($_POST['length1b']);
-	$length1c	= mysql_real_escape_string($_POST['length1c']);
-	$length2	= mysql_real_escape_string($_POST['length2']);
+	$length1a	= mysql_real_escape_string($_POST['length1a']) * 60;
+	$length1b	= mysql_real_escape_string($_POST['length1b']) * 60;
+	$length1c	= mysql_real_escape_string($_POST['length1c']) * 60;
+	$length2	= mysql_real_escape_string($_POST['length2']) * 60;
 	
-	$query = "DELETE FROM `Game` WHERE `Organizer ID` = '".$_SESSION['username']."';" ;
-	$data		= mysql_query($query,$connection);
-	
-	$query		= "INSERT INTO `Game` (`Organizer ID`,`Starting time`, 
-				  `Length 1a`,`Length 1b`,`Length 1c`,`Length 2`, `Starting time Phase 2`)
-				   VALUES ('".$_SESSION['username']."', '2999-12-12 11:11:11' ,".$length1a.
-				   ", ".$length1b.", 
-				   ".$length1c.", 
-				   ".$length2.",
-				   '2999-12-12 11:11:11' );";
-	$data		= mysql_query($query,$connection);
-
-	if ($data == 1) {
-		header("Location: ../organize.php");
-		$_SESSION['gamePhase'] = 1;
-	}
-	else {
+	if( $length1a < 0 or $length1b < 0 or $length1c < 0 or $length2 < 0 )
+	{
 		print "Please check the values!\n";
-		print "<BR><A HREF=newGame.php>Back to insert</A><BR>"	;
+		print "<BR><A HREF=../newGame.php>Back to insert</A><BR>"	;
 	}
+	else {	
+		$query = "DELETE FROM `Game` WHERE `Organizer ID` = '".$_SESSION['username']."';" ;
+		$data		= mysql_query($query,$connection);
+		
+		$query		= "INSERT INTO `Game` (`Organizer ID`,`Starting time`, 
+					  `Length 1a`,`Length 1b`,`Length 1c`,`Length 2`, `Starting time Phase 2`)
+					   VALUES ('".$_SESSION['username']."', '2999-12-12 11:11:11' ,".$length1a.
+					   ", ".$length1b.", 
+					   ".$length1c.", 
+					   ".$length2.",
+					   '2999-12-12 11:11:11' );";
+		$data		= mysql_query($query,$connection);
+		
+		header("Location: ../organize.php");
+	}
+
 	
 }
 else {
